Extraction of some components from FlightForm
